requirements-ci.txt
alembic==1.10.4
apache-airflow==2.3.3
apache-airflow-providers-cncf-kubernetes==5.3.0
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-google==8.6.0
apache-airflow-providers-http==4.3.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-sqlite==3.3.2
connexion==2.14.2
docker==3.7.3
flask==2.2.5
flask-appbuilder==4.1.2
flask-session==0.5.0
freezegun==1.2.0
jsonlines==3.0
pendulum==2.1.2
pytest==7.2.2
sentry-sdk==1.19.0
sqlalchemy<1.5.0
werkzeug==2.2.3
